# Faster constant compares and numbers