--- a/parse_number_test.go
+++ b/parse_number_test.go
@@ -31,8 +31,8 @@ func TestNumberIsValid(t *testing.T) {
 	// From: https://stackoverflow.com/a/13340826
 	var jsonNumberRegexp = regexp.MustCompile(`^-?(?:0|[1-9]\d*)(?:\.\d+)?(?:[eE][+-]?\d+)?$`)
 	isValidNumber := func(s string) bool {
-		tag, _, _ := parseNumber([]byte(s))
-		return tag != TagEnd
+		tag, _ := parseNumber([]byte(s))
+		return tag != 0
 	}
 	validTests := []string{
 		"0",