```
benchmark                                             old ns/op      new ns/op      delta
BenchmarkApache_builds/copy-32                        138493         138857         +0.26%
BenchmarkApache_builds/nocopy-32                      105770         100223         -5.24%
BenchmarkCanada/copy-32                               11324695       11046309       -2.46%
BenchmarkCanada/nocopy-32                             11322561       11141677       -1.60%
BenchmarkCitm_catalog/copy-32                         1560859        1387570        -11.10%
BenchmarkCitm_catalog/nocopy-32                       1202324        1244282        +3.49%
BenchmarkGithub_events/copy-32                        77414          70334          -9.15%
BenchmarkGithub_events/nocopy-32                      62410          58026          -7.02%
BenchmarkGsoc_2018/copy-32                            1259296        1240525        -1.49%
BenchmarkGsoc_2018/nocopy-32                          1020954        1028086        +0.70%
BenchmarkInstruments/copy-32                          336638         285133         -15.30%
BenchmarkInstruments/nocopy-32                        258841         253191         -2.18%
BenchmarkMarine_ik/copy-32                            12680069       12667499       -0.10%
BenchmarkMarine_ik/nocopy-32                          12128601       12184505       +0.46%
BenchmarkMesh/copy-32                                 3785664        3736734        -1.29%
BenchmarkMesh/nocopy-32                               3778827        3691958        -2.30%
BenchmarkMesh_pretty/copy-32                          4296619        4200696        -2.23%
BenchmarkMesh_pretty/nocopy-32                        4293104        4224935        -1.59%
BenchmarkNumbers/copy-32                              724006         720850         -0.44%
BenchmarkNumbers/nocopy-32                            723446         714401         -1.25%
BenchmarkRandom/copy-32                               1069045        974317         -8.86%
BenchmarkRandom/nocopy-32                             725070         720869         -0.58%
BenchmarkTwitter/copy-32                              621589         586054         -5.72%
BenchmarkTwitter/nocopy-32                            467495         450845         -3.56%
BenchmarkTwitterescaped/copy-32                       998509         944640         -5.39%
BenchmarkTwitterescaped/nocopy-32                     861515         852544         -1.04%
BenchmarkUpdate_center/copy-32                        713592         702087         -1.61%
BenchmarkUpdate_center/nocopy-32                      507995         510048         +0.40% 
```